+++ liblzma-rs/src/lzma/lzma2_encoder.rs
@@ -115,19 +115,19 @@ unsafe extern "C" fn mf_read(
 pub const FASTPOS_BITS: c_int = 13;
 #[inline]
 unsafe extern "C" fn get_dist_slot(dist: u32) -> u32 {
-    if dist < (1) << FASTPOS_BITS + (0 as c_int + 0 as c_int * (FASTPOS_BITS - 1 as c_int)) {
+    if dist < (1) << FASTPOS_BITS + (0 + 0 * (FASTPOS_BITS - 1)) {
         return lzma_fastpos[dist as usize] as u32;
     }
-    if dist < (1) << FASTPOS_BITS + (0 as c_int + 1 as c_int * (FASTPOS_BITS - 1 as c_int)) {
-        return (lzma_fastpos[(dist >> 0 + 1 as c_int * (FASTPOS_BITS - 1 as c_int)) as usize]
+    if dist < (1) << FASTPOS_BITS + (0 + 1 * (FASTPOS_BITS - 1)) {
+        return (lzma_fastpos[(dist >> 0 + 1 * (FASTPOS_BITS - 1)) as usize]
             as u32)
             .wrapping_add(
-                (2 as c_int * (0 as c_int + 1 as c_int * (FASTPOS_BITS - 1 as c_int))) as u32,
+                (2 * (0 + 1 * (FASTPOS_BITS - 1))) as u32,
             );
     }
-    return (lzma_fastpos[(dist >> 0 + 2 as c_int * (FASTPOS_BITS - 1 as c_int)) as usize] as u32)
+    return (lzma_fastpos[(dist >> 0 + 2 * (FASTPOS_BITS - 1)) as usize] as u32)
         .wrapping_add(
-            (2 as c_int * (0 as c_int + 2 as c_int * (FASTPOS_BITS - 1 as c_int))) as u32,
+            (2 * (0 + 2 * (FASTPOS_BITS - 1))) as u32,
         );
 }
 pub const LZMA2_CHUNK_MAX: c_uint = 1u32 << 16;
@@ -139,14 +139,14 @@ unsafe extern "C" fn lzma2_header_lzma(coder: *mut lzma_lzma2_coder) {
     if (*coder).need_properties {
         pos = 0;
         if (*coder).need_dictionary_reset {
-            (*coder).buf[pos as usize] = (0x80 as c_int + ((3 as c_int) << 5)) as u8;
+            (*coder).buf[pos as usize] = (0x80 + ((3) << 5)) as u8;
         } else {
-            (*coder).buf[pos as usize] = (0x80 as c_int + ((2 as c_int) << 5)) as u8;
+            (*coder).buf[pos as usize] = (0x80 + ((2) << 5)) as u8;
         }
     } else {
         pos = 1;
         if (*coder).need_state_reset {
-            (*coder).buf[pos as usize] = (0x80 as c_int + ((1 as c_int) << 5)) as u8;
+            (*coder).buf[pos as usize] = (0x80 + ((1) << 5)) as u8;
         } else {
             (*coder).buf[pos as usize] = 0x80 as u8;
         }
